third_party/blink/web_tests/svg/hixie/error/011-expected.txt

CONSOLE ERROR: Error: <svg> attribute viewBox: Expected number, "0 0,, 1 1".
